# tp/concurrency/correction/ramananstsoa-khera-sammandamourthy/tp-modelisation/specifications/spec-swagger.yml
swagger: "2.0"
info:
  version: "1.0.0"
  title: "Swagger Document Edition "
host: "localhost:9090"
basePath: "/api/docs"
schemes:
  - "https"
  - "http"
paths:
  /api/docs:
    post:
      tags:
        - documents
      summary: create a document
      parameters:
        - name: document
          in: body
          required: true
          schema:
            $ref: "#/definitions/docFormCreate"
      responses:
        201:
          description: le document est créé
          schema:
            $ref: '#/definitions/doc'
        400:
          description: le document n'est pas crée
          schema:
            $ref: '#/definitions/errorMessages'
    get:
      tags:
        - documents
      summary: retourne tous les documents, pas de filtrage
      parameters:
        - name: page
          type: integer
          in: query
          description: numéro de la page à retourner
        - name: size
          type: integer
          in: query
          description: nombre de documents par page
      responses:
        200:
          description: la liste des documents
          schema:
            $ref: '#/definitions/docsList'
  /api/docs/{documentId}:
    parameters:
      - name: documentId
        type: string
        in: path
        required: true
        description: identifiant du document
    get:
      summary: lis le document
      tags:
        - documents
      description: retourne un document
      responses:
        200:
          description: le document demandé
          schema:
            $ref: '#/definitions/doc'
        500:
          description: le document n'existe pas
          schema:
            $ref: '#/definitions/errorMessages'
    put:
      tags:
        - documents
      summary: mise à jour du document
      description: |
        met à jour le document si l'utilisateur possède la dernière version et que personne n'a posé de verrou
      parameters:
        - name: document
          in: body
          schema:
            $ref: '#/definitions/docFormUpdate'
          description: met à jour le texte, le titre, l'editeur et la date de mise à jour
      responses:
        200:
          description: le document est mis à jour
          schema:
            $ref: '#/definitions/doc'
        500:
          description: le document ne peut pas être mis à jour car tous les champs sont obligatoires ou que la version n'est pas correcte/renseignée
          schema:
            $ref: '#/definitions/errorMessages'
        409:
          description: le document ne peut pas être mis à jour car un lock existe
          schema:
            $ref: '#/definitions/errorMessages'
  /api/docs/{documentId}/lock:
    parameters:
      - name: documentId
        type: string
        in: path
        required: true
        description: identifiant du document
    get:
      tags:
        - locks
      summary: retourne le verrou posé sur le document si présent
      responses:
        200:
          description: le verrou posé
          schema:
            $ref: '#/definitions/lock'
        204:
          description: aucun verrou posé
    put:
      tags:
        - locks
      summary: pose un verrou sur le document
      description: |
        l'utilisateur peut poser un verrou si aucun n'autre n'est posé
      parameters:
        - name: lock
          in: body
          description: l'objet verrou posé
          schema:
            $ref: "#/definitions/lockFormCreate"
      responses:
        200:
          description: le verrou est posé
          schema:
            $ref: "#/definitions/lock"
        409:
          description: un verrou est déjà posé, retourne le verrou déjà posé
          schema:
            $ref: "#/definitions/lock"
    delete:
      tags:
        - locks
      summary: supprime le verrou posé
      responses:
        204:
          description: le verrou est supprimé
definitions:
  pageData:
    description: objet générique pour les résultats de recherche
    type: object
    properties:
      totalElements:
        description: le nombre total d'element
        type: integer
      page:
        type: integer
        description: le numéro de la page courante
  docFormCreate:
    description: créer un document
    type: object
    properties:
      title:
        type: string
        description: titre du document
      creator:
        type: string
        description: le nom du créateur du document
      editor:
        type: string
        description: le nom du dernier utilisateur qui l'a mis à jour
      body:
        type: string
        description: le texte du document
  docFormUpdate:
    description: modifier un document
    type: object
    properties:
      title:
        type: string
        description: titre du document
      creator:
        type: string
        description: le nom du créateur du document
      editor:
        type: string
        description: le nom du dernier utilisateur qui l'a mis à jour
      body:
        type: string
        description: le texte du document
      version:
        type: string
        description: version du document
  docSummary:
    description: résumé d'un document
    type: object
    properties:
      customLock:
        type: string
        description: document bloqué ou non
      documentId:
        type: string
        description: identifiant du document
      created:
        type: string
        format: date-time
        description: la date de création
      updated:
        type: string
        format: date-time
        description: date de la mise à jour
      title:
        type: string
        description: titre du document
      creator:
        type: string
        description: le nom du créateur du document
      editor:
        type: string
        description: le nom du dernier utilisateur qui l'a mis à jour
      body:
        type: string
        description: le texte du document
      version:
        type: string
        description: version du document
  doc:
    description: a document
    allOf:
      - $ref: "#/definitions/docSummary"
    properties:
      creator:
        type: string
        description: le nom du créateur du document
      editor:
        type: string
        description: le nom du dernier utilisateur qui l'a mis à jour
      body:
        type: string
        description: le texte du document
      version:
        type: integer
        description: la version du document
  docsList:
    description: la liste des résultats d'une recherche de document
    allOf:
      - $ref: '#/definitions/pageData'
    properties:
      data:
        type: array
        items:
          $ref: "#/definitions/docSummary"
  lock:
    description: un verrou
    type: object
    properties:
      id:
        type: string
        description: id du verrou
      owner:
        type: string
        description: utilisateur propriétaire du verrou
      created:
        description: date de la pose du verrou
        type: string
        format: date-time
  lockFormCreate:
    description: un verrou
    type: object
    properties:
      owner:
        type: string
        description: utilisateur propriétaire du verrou
  errorMessages:
    description: la description de l'erreur
    type: object
    properties:
      errorMessageType:
        type: array
        items:
          $ref: "#/definitions/errorMessageType"
      errors:
        type: array
        items:
          type: object
          properties:
            errorCode:
              type: string
              description: le code de l'erreur
            errorMessage:
              type: string
              description: message d'erreur
  errorMessage:
    description: la description de l'erreur
    type: object
    properties:
      errorCode:
        type: string
        description: code HTTP status
      errorMesage:
        type: string
        description: la description du message
  errorMessageType:
    description: le type d'erreur
    type: object
    properties:
      type:
        type: string
        description: type d'erreur
        enum:
          - TECHNICAL
          - FUNCTIONAL